edit order list

+++ dash-pages/order.php
@@ -22,16 +22,20 @@ if (isset($_SESSION) && !empty($_SESSION['client'])) {
                 <?php
                 require_once 'method/order.php';
                 $orderList = order::orderListByidClient($id_client);
+                if ($orderList != null) {
+
+
                 ?>
-                <?php foreach ($orderList as $a) { ?>
-                    <tr>
-                        <td><?= $a['id_commande'] ?></td>
-                        <td><?= $a['nom'] ?></td>
-                        <td>$<?= $a['prix'] ?></td>
-                        <td><?= $a['category'] ?></td>
-                        <td><a href="action/delete.php?key=com&id=<?= $a['id_commande'] ?>"> <i style="color:red" class="far fa-trash-alt"></i></a></td>
-                    </tr>
-                <?php } ?>
+                    <?php foreach ($orderList as $a) { ?>
+                        <tr>
+                            <td><?= $a['id_commande'] ?></td>
+                            <td><?= $a['nom'] ?></td>
+                            <td>$<?= $a['prix'] ?></td>
+                            <td><?= $a['category'] ?></td>
+                            <td><a href="action/delete.php?key=com&id=<?= $a['id_commande'] ?>"> <i style="color:red" class="far fa-trash-alt"></i></a></td>
+                        </tr>
+                <?php }
+                } ?>
             </tbody>
 
         </table>
